Refactor

+++ template.rs
@@ -4,7 +4,7 @@ fn read<T: std::str::FromStr>() -> T {
   buf.trim().parse::<T>().ok().unwrap()
 }
 
-fn reads<T: std::str::FromStr>(n: u32) -> Vec<T> {
+fn reads<T: std::str::FromStr>(n: usize) -> Vec<T> {
   let mut vec: Vec<T> = vec![];
   for _ in 0 .. n {
     vec.push(read());
@@ -21,7 +21,7 @@ fn read_tuple<T1: std::str::FromStr, T2: std::str::FromStr>() -> (T1, T2) {
   (x, y)
 }
 
-fn read_tuples<T1: std::str::FromStr, T2: std::str::FromStr>(n: u32) -> Vec<(T1, T2)> {
+fn read_tuples<T1: std::str::FromStr, T2: std::str::FromStr>(n: usize) -> Vec<(T1, T2)> {
   let mut vec: Vec<(T1, T2)> = vec![];
   for _ in 0 .. n {
     vec.push(read_tuple());
